+++ pydbml/exceptions.py
@@ -1,0 +1,10 @@
+class TableNotFoundError(Exception):
+    pass
+
+
+class ColumnNotFoundError(Exception):
+    pass
+
+
+class AttributeMissingError(Exception):
+    pass