+++ pydbml/parser.py
@@ -8,20 +8,29 @@ from pydbml.definitions.enum import enum
 from pydbml.definitions.table_group import table_group
 from pydbml.definitions.project import project
 from pydbml.classes import Table, TableReference, Reference
-
+from pydbml.exceptions import TableNotFoundError, ColumnNotFoundError
 
 pp.ParserElement.setDefaultWhitespaceChars(' \t\r')
 
 
-class TableNotFoundError(Exception):
-    pass
-
-
-class ColumnNotFoundError(Exception):
-    pass
-
-
 class PyDBML:
+    '''
+    PyDBML parser factory. If properly initiated, returns PyDBMLParseResults
+    which contains parse results in attributes.
+
+    Usage option 1:
+
+    >>> with open('schema.dbml') as f:
+    ...     p = PyDBML(f)
+    ...     # or
+    ...     p = PyDBML(f.read())
+
+    Usage option 2:
+    >>> p = PyDBML.parse_file('schema.dbml')
+    >>> # or
+    >>> from pathlib import Path
+    >>> p = PyDBML(Path('schema.dbml'))
+    '''
     def __new__(cls,
                 source_: str or PosixPath or TextIOWrapper or None = None):
         if source_ is not None:
@@ -58,6 +67,7 @@ class PyDBMLParseResults:
         self.tables = []
         self.table_dict = {}
         self.refs = []
+        self.ref_blueprints = []
         self.enums = []
         self.table_groups = []
         self.project = None
@@ -66,7 +76,7 @@ class PyDBMLParseResults:
         self._set_syntax()
         self._syntax.parseString(self.source, parseAll=True)
         self._validate()
-        self._add_table_refs()
+        self._process_refs()
         self._set_enum_types()
 
     def __repr__(self):
@@ -80,7 +90,7 @@ class PyDBMLParseResults:
         project_expr = project.copy()
 
         table_expr.addParseAction(self._parse_table)
-        ref_expr.addParseAction(self._parse_ref)
+        ref_expr.addParseAction(self._parse_ref_blueprint)
         enum_expr.addParseAction(self._parse_enum)
         table_group_expr.addParseAction(self._parse_table_group)
         project_expr.addParseAction(self._parse_project)
@@ -107,11 +117,11 @@ class PyDBMLParseResults:
         table = t[0]
         self.tables.append(table)
         for col in table.columns:
-            self.refs.extend(col.refs)
+            self.ref_blueprints.extend(col.ref_blueprints)
         self.table_dict[table.name] = table
 
-    def _parse_ref(self, s, l, t):
-        self.refs.append(t[0])
+    def _parse_ref_blueprint(self, s, l, t):
+        self.ref_blueprints.append(t[0])
 
     def _parse_enum(self, s, l, t):
         self.enums.append(t[0])
@@ -125,30 +135,64 @@ class PyDBMLParseResults:
         else:
             raise SyntaxError('Project redifinition not allowed')
 
-    def _add_table_refs(self):
-        for ref_ in self.refs:
+    def _process_refs(self):
+        '''
+        Fill up the `refs` attribute with Reference object, created from
+        reference blueprints;
+        Add TableReference objects to each table which has references.
+        Validate refs at the same time.
+        '''
+        for ref_ in self.ref_blueprints:
+            table1 = self.table_dict.get(ref_.table1)
+            if not table1:
+                raise TableNotFoundError('Error while parsing reference:'
+                                         f'table "{ref_.table1}"" is not defined.')
+            table2 = self.table_dict.get(ref_.table2)
+            if not table2:
+                raise TableNotFoundError('Error while parsing reference:'
+                                         f'table "{ref_.table2}"" is not defined.')
+            col1 = table1.get(ref_.col1)
+            if not col1:
+                raise ColumnNotFoundError('Error while parsing reference:'
+                                          f'column "{ref_.col1} not defined in table "{table1.name}".')
+            col2 = table2.get(ref_.col2)
+            if not col2:
+                raise ColumnNotFoundError('Error while parsing reference:'
+                                          f'column "{ref_.col2} not defined in table "{table2.name}".')
+            self.refs.append(
+                Reference(
+                    ref_.type,
+                    table1,
+                    col1,
+                    table2,
+                    col2,
+                    name=ref_.name,
+                    comment=ref_.comment,
+                    on_update=ref_.on_delete,
+                    on_delete=ref_.on_update
+                )
+            )
+
             if ref_.type in (Reference.MANY_TO_ONE, Reference.ONE_TO_ONE):
-                table = self.table_dict[ref_.table1]
+                table = table1
                 init_dict = {
-                    'col': ref_.col1,
-                    'ref_table': ref_.table2,
-                    'ref_col': ref_.col2,
+                    'col': col1,
+                    'ref_table': table2,
+                    'ref_col': col2,
+                    'name': ref_.name,
                     'on_update': ref_.on_update,
                     'on_delete': ref_.on_delete
                 }
-                if ref_.name:
-                    init_dict['name'] = ref_.name
             else:
-                table = self.table_dict[ref_.table2]
+                table = table2
                 init_dict = {
-                    'col': ref_.col2,
-                    'ref_table': ref_.table1,
-                    'ref_col': ref_.col1,
+                    'col': col2,
+                    'ref_table': table1,
+                    'ref_col': col1,
+                    'name': ref_.name,
                     'on_update': ref_.on_update,
                     'on_delete': ref_.on_delete
                 }
-                if ref_.name:
-                    init_dict['name'] = ref_.name
             table.refs.append(
                 TableReference(**init_dict)
             )
@@ -161,28 +205,20 @@ class PyDBMLParseResults:
                     col.type = enum_dict[str(col.type)].get_type()
 
     def _validate(self):
-        for ref_ in self.refs:
-            try:
-                table1 = self.table_dict[ref_.table1]
-            except KeyError:
-                raise TableNotFoundError(f"Can't find table \"{ref_.table1}\"")
-            for col in table1:
-                if ref_.col1 == col.name:
-                    break
-            else:
-                raise ColumnNotFoundError(f'Can\'t find column "{ref_.col1}" in table "{ref_.table1}"')
+        self._validate_table_groups()
 
-            try:
-                table2 = self.table_dict[ref_.table2]
-            except KeyError:
-                raise TableNotFoundError(f"Can't find table \"{ref_.table2}\"")
-            for col in table2:
-                if ref_.col2 == col.name:
-                    break
-            else:
-                raise ColumnNotFoundError(f'Can\'t find column "{ref_.col2}" in table "{ref_.table2}"')
+    def _validate_table_groups(self):
+        '''
+        Check that all tables, mentioned in the table groups, exist
+        '''
+        for tg in self.table_groups:
+            for table_name in tg:
+                if table_name not in self.table_dict:
+                    raise TableNotFoundError(f'Cannot add Table Group "{tg.name}": table "{table_name}" not found.')
 
     @property
     def sql(self):
+        '''Returs SQL of the parsed results'''
+
         components = (i.sql for i in (*self.enums, *self.tables))
         return '\n'.join(components)